side-projects/intelligent-calendar-agent/backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.27
openai>=1.40
google-api-python-client>=2.130
google-auth-oauthlib>=1.2
google-auth-httplib2>=0.2
pydantic>=2.6
pydantic-settings>=2.2
pytest>=8.0
